python/devtools/test_eip712_signing.py
#!/usr/bin/env python3
"""
Test EIP-712 Signing with CDP Signer

This script tests the CDP signer's EIP-712 signing capabilities
using the native CDP SDK methods.
"""

import asyncio
import json
import sys
from pathlib import Path

# Add src to path
sys.path.insert(0, str(Path(__file__).parent / "src"))

from client.core.cdp_signer import create_cdp_signer
from shared.utils.logger import logger

async def test_eip712_signing():
    """Test EIP-712 signing with CDP signer"""
    
    print("🧪 Testing EIP-712 Signing with CDP Signer")
    print("=" * 50)
    
    try:
        # Create CDP signer
        print("1. Creating CDP signer...")
        cdp_signer = await create_cdp_signer("test-client")
        print(f"   ✅ CDP signer created")
        print(f"   • Address: {cdp_signer.address}")
        print(f"   • Account Type: {type(cdp_signer.account).__name__}")
        print()
        
        # Test EIP-712 domain and message
        print("2. Testing EIP-712 signing...")
        
        # X402-style domain
        domain = {
            "name": "X402 Payment",
            "version": "1",
            "chainId": 84532,  # Base Sepolia
            "verifyingContract": "0x0000000000000000000000000000000000000000"
        }
        
        # X402-style types (with EIP712Domain definition)
        types = {
            "EIP712Domain": [
                {"name": "name", "type": "string"},
                {"name": "version", "type": "string"},
                {"name": "chainId", "type": "uint256"},
                {"name": "verifyingContract", "type": "address"}
            ],
            "Payment": [
                {"name": "amount", "type": "uint256"},
                {"name": "recipient", "type": "address"},
                {"name": "nonce", "type": "uint256"},
                {"name": "deadline", "type": "uint256"}
            ]
        }
        
        # Test message
        message = {
            "amount": "1000000",  # 1 USDC in wei
            "recipient": "0x742d35Cc6634C0532925a3b8D4C9db96C4b4d8b6",
            "nonce": "12345",
            "deadline": "1704067200"  # Some future timestamp
        }
        
        print("   • Domain:", json.dumps(domain, indent=2))
        print("   • Types:", json.dumps(types, indent=2))
        print("   • Message:", json.dumps(message, indent=2))
        print()
        
        # Sign the typed data
        print("3. Signing typed data...")
        signature = await cdp_signer.sign_typed_data(domain, types, "Payment", message)
        
        print(f"   ✅ EIP-712 signature generated!")
        print(f"   • Signature: {signature}")
        print(f"   • Length: {len(signature)} characters")
        print()
        
        # Verify signature format (should be 0x + 130 hex chars)
        if signature.startswith("0x") and len(signature) == 132:
            print("   ✅ Signature format is correct (0x + 130 hex chars)")
        else:
            print("   ⚠️  Signature format may be unexpected")
        print()
        
        # Test other signing methods
        print("4. Testing other signing methods...")
        
        # Test sign_message (EIP-191)
        try:
            from eth_account.messages import encode_defunct
            from eth_utils import text_if_str, to_bytes
            
            message_text = "Hello X402!"
            message_hash = encode_defunct(text_if_str(to_bytes, message_text))
            signed_message = cdp_signer.sign_message(message_hash)
            
            print(f"   ✅ EIP-191 message signing works")
            print(f"   • Message: {message_text}")
            print(f"   • Signature: {signed_message.signature.hex()}")
        except Exception as e:
            print(f"   ❌ EIP-191 message signing failed: {e}")
        
        print()
        
        # Test sign_transaction
        try:
            transaction = {
                "to": "0x742d35Cc6634C0532925a3b8D4C9db96C4b4d8b6",
                "value": 1000000000000000,  # 0.001 ETH
                "gas": 21000,
                "gasPrice": 20000000000,  # 20 gwei
                "nonce": 0,
                "chainId": 84532
            }
            
            signed_tx = cdp_signer.sign_transaction(transaction)
            print(f"   ✅ Transaction signing works")
            print(f"   • Raw transaction: {signed_tx.rawTransaction.hex()}")
        except Exception as e:
            print(f"   ❌ Transaction signing failed: {e}")
        
        print()
        
        print("🎉 CDP Signer EIP-712 Test Complete!")
        print("=" * 50)
        print("✅ All signing methods are working")
        print("✅ Ready for X402 integration")
        print("✅ CDP SDK provides full signing capabilities")
        
    except Exception as e:
        print(f"❌ Test failed: {e}")
        logger.error("EIP-712 signing test failed", e)

if __name__ == "__main__":
    asyncio.run(test_eip712_signing()) 