python/requirements.txt
cdp-sdk
click
python-dotenv
rich
typer
pydantic
fastapi
uvicorn[standard]
requests
x402
eth-account 